+++ R/guiutils.R
@@ -1,24 +1,24 @@
 ## Copyright (c) 2004-2015, Ph. Grosjean <phgrosjean@sciviews.org>
 ##
 ## This file is part of ZooImage
-## 
+##
 ## ZooImage is free software: you can redistribute it and/or modify
 ## it under the terms of the GNU General Public License as published by
 ## the Free Software Foundation, either version 2 of the License, or
 ## (at your option) any later version.
-## 
+##
 ## ZooImage is distributed in the hope that it will be useful,
 ## but WITHOUT ANY WARRANTY; without even the implied warranty of
 ## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 ## GNU General Public License for more details.
-## 
+##
 ## You should have received a copy of the GNU General Public License
 ## along with ZooImage. If not, see <http://www.gnu.org/licenses/>.
 
 ## Get the name of one or several objects of a given class
 selectObject <- function (class = "data.frame", default = "", multiple = FALSE,
 title = paste0("Choose a ", class, ":"))
-{	
+{
 	objlist <- ls(envir = .GlobalEnv,)	# Get objects in .GlobalEnv
 	if (!length(objlist)) {
 		warning("There is no object of class '", paste(class, collapse = " "),
@@ -30,9 +30,9 @@ title = paste0("Choose a ", class, ":"))
 	for (i in 1:length(objlist))
 		Filter[i] <- inherits(get(objlist[i], envir = .GlobalEnv,
 			inherits = FALSE), class)
-	
+
 	## Keep only those objects
-	objlist <- objlist[Filter]	
+	objlist <- objlist[Filter]
 	if (!length(objlist)) {	# No such objects in .GlobalEnv
 		warning("There is no object of class '", paste(class, collapse = " "),
 			"' in the user workspace!")
@@ -40,32 +40,32 @@ title = paste0("Choose a ", class, ":"))
 	}
 	if (default == "") default <- objlist[1]
 	dlgList(objlist, preselect = default, multiple = multiple,
-		title = title)$res		
+		title = title)$res
 }
 
 ## Get the name of one or more lists with their components of a given class
 selectList <- function (class = "data.frame", default = "", multiple = FALSE,
 title = paste0("Choose a list (of ", class, "s):"))
-{	
+{
 	filter <- function (x) {
 		item <- get(x, envir = .GlobalEnv, inherits = FALSE)
 		is.list(item) && all(sapply(item, function (y) inherits(y, class)))
 	}
-	varlist <- Filter(filter, objects(pos = 1))	
+	varlist <- Filter(filter, objects(pos = 1))
 	if (length(varlist) == 0) {
 		warning("no list of '", class, "' objects in the user workspace")
 		return(character(0))
 	}
 	if (default == "") default <- varlist[1]
 	dlgList(varlist, preselect = default, multiple = multiple,
-		title = title)$res		
+		title = title)$res
 }
 
 ## Select one or several files of a given type
 selectFile <- function (type = c("ZipZid", "ZimZis", "LstZid", "ZidZidb",
-"Zip", "Zid", "Zidb", "Zim", "Zis", "Zie", "Zic", "Img", "TifPgm", "RData"),
+"Zip", "Zid", "Zidb", "Zim", "Zis", "Zie", "Zic", "Img", "TifPgm", "RData", "R"),
 multiple = FALSE, quote = TRUE, title = NULL)
-{	
+{
 	type <- match.arg(type)
 	Type <- switch(type,
 		ZipZis = "Zip/Zis",
@@ -74,39 +74,40 @@ multiple = FALSE, quote = TRUE, title = NULL)
 		TifPgm = "Tiff/Pgm",
 		ZidZidb = "Zid/Zidb",
 		type)
-	
+
 	## Adapt title according to 'multiple'
 	if (isTRUE(as.logical(multiple)) && !is.null(title)) {
 		title <- paste("Select one or several", Type, "files...")
 	} else title <- paste("Select one", Type, "file...")
-	
+
 	filters <- switch(type,
-		ZipZid 	= c("ZooImage files"          		, ".zip",
-					"ZooImage files"          		, ".zid"),
-		ZimZis 	= c("ZooImage metadata files" 		, ".zim",
-					"ZooImage metadata files" 		, ".zis"),
-		LstZid  = c("FlowCAM list files"      		, ".lst",
-					"ZooImage files"          		, ".zid"),
-		ZidZidb = c("ZooImage files"          		, ".zid",
-					"ZooImage databases"      		, ".zidb"),
-		Zip		= c("ZooImage picture files"  		, ".zip"),
-		Zid		= c("ZooImage data files"     		, ".zid"),
-		Zidb    = c("ZooImage databases"      		, ".zidb"),
-		Zim		= c("ZooImage metadata files" 		, ".zim"),
-		Zis		= c("ZooImage sample files"   		, ".zis"),
-		Zie		= c("ZooImage extension files"		, ".zie"),
-		Zic     = c("ZooImage Classification Scheme",".zic" ),
-		Img     = c("Tiff image files"        		, ".tif",
-					"Jpeg image files"        		, ".jpg",
-					"Zooimage import extensions"	,".zie",
-					"Table and ImportTemplate.zie"	,".txt",
-					"FlowCAM Table and ImportTemplate.zie",".txt"),
-		TifPgm  = c("Tiff image files"        		, ".tif",
-					"Pgm image files"         		, ".pgm"),
-		RData   = c("R data"                  		, ".RData"))
+		ZipZid 	= c("ZooImage files (*.zid)"                , "*.zid",
+					"ZooImage files (*.zip)"                      , "*.zid"),
+		ZimZis 	= c("ZooImage metadata files (*.zim)"       , "*.zim",
+					"ZooImage metadata files (*.zis)"             , "*.zis"),
+		LstZid  = c("FlowCAM list files (*.lst)"            , "*.lst",
+					"ZooImage files (*.zid)"                      , "*.zid"),
+		ZidZidb = c("ZooImage databases (*.zidb)"           , "*.zidd",
+					"ZooImage files (*.zid)"                      , "*.zid"),
+		Zip		= c("ZooImage picture files (*.zip)"          , "*.zip"),
+		Zid		= c("ZooImage data files (*.zid)"             , "*.zid"),
+		Zidb    = c("ZooImage databases (*.zidb)"           , "*.zidb"),
+		Zim		= c("ZooImage metadata files (*.zim)"         , "*.zim"),
+		Zis		= c("ZooImage sample files (*.zis)"           , "*.zis"),
+		Zie		= c("ZooImage extension files (*.zie)"        , "*.zie"),
+		Zic     = c("ZooImage Classification Scheme (*.zic)", "*.zic" ),
+		Img     = c("Tiff image files (*.tif)"              , "*.tif",
+					"Jpeg image files (*.jpg)"                    , "*.jpg",
+					"Zooimage import extensions (*.zie)"          , ".*zie",
+					"Table and ImportTemplate.zie (*.txt)"        , "*.txt",
+					"FlowCAM Table and ImportTemplate.zie (*.txt)", "*.txt"),
+		TifPgm  = c("Tiff image files (*.tif)"              , "*.tif",
+					"Pgm image files (*.pgm)"                     , "*.pgm"),
+		RData   = c("R data (*.Rdata)"                      , "*.RData"),
+	  R       = c("R script (*.R)"                        , "*.R"))
 	filters <- matrix(filters, ncol = 2, byrow = TRUE)
-	
-	res <- dlgOpen(title = title, multiple = multiple, filters = filters)$res	
+
+	res <- dlgOpen(title = title, multiple = multiple, filters = filters)$res
 	if (length(res) && res != "" && quote)
 		res <- paste('"', res, '"', sep = "")
 	res
@@ -144,7 +145,7 @@ imageViewer <- function (dir = getwd(), pgm = getOption("ZI.ImageViewer"))
 			dir)
 		system(cmd, wait = FALSE, ignore.stdout = TRUE, ignore.stderr = TRUE)
 	} else {
-		cmd <- sprintf('nautilus --geometry 600x600 "%s"', dir) 
+		cmd <- sprintf('nautilus --geometry 600x600 "%s"', dir)
 		system(cmd, wait = FALSE, ignore.stdout = TRUE, ignore.stderr = TRUE)
 	}
 }
@@ -154,7 +155,7 @@ iconize = FALSE, wait = FALSE)
 {
 	## Look if the program path is recorded in the options
 	pgmPath <- getOption(program)
-	if (!is.null(pgmPath) && file.exists(pgmPath)) {
+	if (!is.null(pgmPath) && pgmPath != "") { #> no -> && file.exists(pgmPath)) {
 		## Do we need to switch directory?
 		if (switchdir) {
 			curdir <- setwd(dirname(pgmPath))
@@ -169,13 +170,15 @@ iconize = FALSE, wait = FALSE)
 #		tkwm.iconify(WinGet("ZIDlgWin"))
 }
 
+fileEdit <- svMisc::fileEdit
+
 modalAssistant <- function (title, text, init, options = NULL, check = NULL,
 select.file = NULL, returnValOnCancel = "ID_CANCEL", help.topic = NULL)
 {
 	## TODO!!!!
 	message("Modal assistant temporarily disabled!")
 	return(returnValOnCancel)
-	
+
 #	## Create an assistant dialog box which behaves as a modal dialog box
 #	text <- paste(text, collapse = "\n")
 #	try(tkWinAdd("ZIAssist", title = title, bind.delete = FALSE))